refactor(influxql): remove `copy_with_replacement` and use upstream API (#8451)
* refactor(influxql): remove `copy_with_replacement` and use upstream version

* fix: clippy

---------
