+++ iox_query_influxql/src/plan/udf.rs
@@ -5,7 +5,7 @@
 //! call information as the InfluxQL AST. These expressions are then
 //! rewritten at a later stage of planning, with more context available.
 
-use crate::plan::util_copy::find_exprs_in_exprs;
+use crate::plan::util::find_exprs_in_exprs;
 use crate::{error, NUMERICS};
 use arrow::datatypes::{DataType, TimeUnit};
 use datafusion::logical_expr::{
